Move out validationUtils from user package
